// src/types/post.ts
export interface ICreateComments {
  text: Descendant[];
}
// ...
export interface ISendPost {
  audience: Audience;
  mediaUrls: any[];
  text: Descendant[];
  editMode: boolean;
  id?: string;
  postType?: string;
}